# writing reflection <--! great read -->
In my opinion, there could be a few different forms of contemporary tools that could be used to create art that is a Gesamtkunstwerk, or a full work of art made of multiple components. When one looks at artwork, some elements may pop out more than others. Some may be less visible, or less eye catching than others, but each serves its own purpose. One thing I feel could be used is dramatic lighting. Dramatic light can be shown on the side of a person's face, their body, or even their entire face. This can help the audience feel the element of surprise or suspense leading up to what the actor is supposed to portray in a play. I also believe that as aforementioned in the Entangled reading by Chris Salter, sound affects are very key in conveying different things as well, Depending on how loud or soft the music is, one can also feel certain emotions of intensity, or serenity in a given setting, or point in a play.


      Another thing that I believe can really sell a Gesamtkunstwerk is a collection of varying facial expressions. As we've often heard, one's body language, or facial expressions can say a lot about a person, or how they're feeling toward the speaker/audience. Some prime examples from Net Art Anthology are Arcangel Surfware, The World in 24 Hours, and Floodnet. Varying facial expressions in one scene can show the multiple emotions faced by the different characters in the room at any given time. It makes for a much more entertaining performance, and overall visual appeal within stage acting. Lastly, the atmosphere, or overall visual scene can play a big role in how a Gesamtkunstwerkis decided upon. Often times, when we think of a beautiful scenery, we consider it a complete work of art when we can see it very well from an intriguing angle.  For example, in Net Art Anthology, The Thing, Untitled, and Starrynight do a great job conveying this example. When one is able to see a wide spectrum, or field of view, it cause cause a dramatic feeling, and visual affect. As similarly mentioned in the article, when the stage came in one show formed a pyramid like shape, and everything came together as one, it made for a great spectacle for the eye to see, and a much more enjoyable atmosphere and performance.